Added log parsing to build
