+++ src/handlers/events/delivery/deploy/Deployer.ts
@@ -2,23 +2,9 @@ import { RemoteRepoRef } from "@atomist/automation-client/operations/common/Repo
 import { DeployableArtifact } from "../ArtifactStore";
 import { QueryableProgressLog } from "../log/ProgressLog";
 import { Deployment, TargetInfo } from "./Deployment";
+import { LogInterpretation } from "../log/InterpretedLog";
 
-export interface InterpretedLog {
-
-    /**
-     * Relevant part of log to display in UX, if we were able to identify it
-     */
-    relevantPart: string;
-
-    message: string;
-
-    /**
-     * Should the UX include the full log, or is it too long or ugly?
-     */
-    includeFullLog?: boolean;
-}
-
-export interface Deployer<T extends TargetInfo = TargetInfo> {
+export interface Deployer<T extends TargetInfo = TargetInfo> extends LogInterpretation {
 
     /**
      * Implemented by deployers that don't sit on an infrastructure like Cloud Foundry
@@ -28,7 +14,5 @@ export interface Deployer<T extends TargetInfo = TargetInfo> {
     undeploy?(id: RemoteRepoRef): Promise<any>;
 
     deploy(ai: DeployableArtifact, ti: T, log: QueryableProgressLog): Promise<Deployment>;
-
-    errorParser?(log: string): InterpretedLog | undefined;
 
 }
